Add stories frontend
