+++ backend/src/main/java/com/example/backend/controller/PostController.java
@@ -91,16 +91,4 @@ public class PostController {
             throws InterruptedException, ExecutionException, IllegalAccessException {
         return postService.incrementPost("savedPosts", data.getPostId(), data.getUserId(), false);
     }
-
-
-    @PostMapping(value = "/posts/{id}/save")
-    public String savePost(@RequestBody PostUserJoin data) throws InterruptedException, ExecutionException {
-        return postService.savePost(data.getPostId(), data.getUserId(), true);
-    }
-
-    @DeleteMapping(value = "/posts/{id}/save")
-    public String unsavePost(@RequestBody PostUserJoin data)
-            throws InterruptedException, ExecutionException {
-        return postService.savePost(data.getPostId(), data.getUserId(), false);
-    }
 }
